Lab12 release v14

+++ CMS/cms/comments.php
@@ -3,40 +3,43 @@ include "../inc/connect-inc.php";
 include "../classes/MyDB.php";
 include "classes/ProductModel.php";
 include_once "../inc/filter_input_.php";
+include "classes/UserSessions.php";
 
-//TODO - check session
-
-$error_message = null;
-$mysqli = MyDB::get_db_instance();
-$action = filter_input_("action", "");
-$viewMode = "";
-$model = new ProductModel($mysqli);
-$error_message = null;
-switch ($action) {
-    case "delete":
-        $id = filter_input_("id", 0);
-        ($id != 0) ?
-            $model->deleteProductReviews($id) :
-            $error_message = "Can not delete comment, incorrect id";
-        break;
-    case "add":
-        $id_prod = filter_input_("id_prod", "");
-        $email = filter_input_("email", "");
-        $name = filter_input_("name", "");
-        $comment = filter_input_("comment", "");
-        if ($model->getProduct($id_prod) != 0) {
-            $model->addProductReviews($email, $id_prod, $name, $comment);
-            $id_prod = "";
-            $email = "";
-            $name = "";
-            $comment = "";
-        } else
-            $error_message = "Can not delete comment, no product with this ID";
+$u = new UserSessions();
+if ($u->checkUserAuth() != 0) {
+    $error_message = null;
+    $mysqli = MyDB::get_db_instance();
+    $action = filter_input_("action", "");
+    $viewMode = "";
+    $model = new ProductModel($mysqli);
+    $error_message = null;
+    switch ($action) {
+        case "delete":
+            $id = filter_input_("id", 0);
+            ($id != 0) ?
+                $model->deleteProductReviews($id) :
+                $error_message = "Can not delete comment, incorrect id";
+            break;
+        case "add":
+            $id_prod = filter_input_("id_prod", "");
+            $email = filter_input_("email", "");
+            $name = filter_input_("name", "");
+            $comment = filter_input_("comment", "");
+            if ($model->getProduct($id_prod) != 0) {
+                $model->addProductReviews($email, $id_prod, $name, $comment);
+                $id_prod = "";
+                $email = "";
+                $name = "";
+                $comment = "";
+            } else
+                $error_message = "Can not delete comment, no product with this ID";
+    }
+    if ($viewMode == "")
+        $list = $model->getProductsReviews();
+    $mysqli->close();
+} else {
+    header('Location: http://k503labs.ukrdomen.com/535a/Veprytskaya/CMS/cms/index.php');
 }
-
-if ($viewMode == "")
-    $list = $model->getProductsReviews();
-$mysqli->close();
 include "inc/header.php"; ?>
     <table id="customers">
         <tr>
